Bug fixing constant dtype

+++ ssl_image_classification/main.py
@@ -198,8 +198,8 @@ def main(parser_args=[], console_args=False):
         test_writer = tf.summary.create_file_writer(f"{log_path}/test")
 
     # Assigning args used in functions wrapped with tf.function to tf.constant/tf.Variable to avoid memory leaks
-    args["T"] = tf.constant(args["T"])
-    args["K"] = tf.constant(args["K"])
+    args["T"] = tf.constant(args["T"], dtype=tf.float32)
+    args["K"] = tf.constant(args["K"], dtype=tf.float32)
     args["epochs"] = tf.constant(args["epochs"])
     args["pre_val_iter"] = tf.constant(args["pre_val_iter"])
     args["threshold"] = tf.constant(args["threshold"])
